added type casting in avg

+++ src/phitron conceptual/C/sum_of_three_num.c
@@ -5,5 +5,5 @@ int main()
     printf("Enter Three Whole Number: ");
     scanf("%d%d%d", &a, &b, &c);
     printf("Sum of %d + %d + %d = %d\n", a, b, c, a + b + c);
-    printf("Average of (%d + %d + %d)/3 = %d", a, b, c, (a + b + c )/ 3);
+    printf("Average of (%d + %d + %d)/3 = %f", a, b, c, (float)(a + b + c )/ 3);
 }
